+++ Makefile
@@ -3,9 +3,8 @@ NAME= webserv
 CXX= c++
 CXXFLAGS= -Wall -Wextra -Werror -std=c++98 -Iheaders
 
-# SRC= webserv.cpp HttpRequest.cpp 
+SRC= webserv.cpp HttpRequest.cpp 
 # SRC+= main.cpp Logger.cpp
-SRC = main.cpp new_parser.cpp
 
 VPATH= src/:obj/:headers/:http_request_parser/
